Safari fixed: Special characters in email-to input
Not only command+a bugged in the email-to from the upload form, but all not character type keystrokes gave problems.

This is due to how the ngx-chips package handles keystrokes, in a very cumbersome way. The reason is to allow functions which are not interesting for us and therefore can be avoided.