Lazy loading modules & components & preloading
Few big structural changes: splitting up the app in two major modules:
- Home module
- Project module

Also made a component module for advertisement which are also lazy loaded. Therefore the sma is now a completely different component.

Result from all this: main.js is now much lower (400KB instead 1.2 MB)

Also preloading fonts now, therefore I did make the fonts static in the assets folder, so they are not hashed anymoreand have static names, which allows the browser to cash these fonts. For this the fonts folder was removed and the assets were moved to the assets folder (they are directly copied to the dist folder when building).